fix: 将 Skill 商店移到技能管理页，修复图标显示
- 移除侧边栏 SkillHub 按钮，改为在技能管理页 (SkillsTab) 新增
  第三个 tab 「SkillHub」，与已有本地技能和 prompts.chat 商店并列
- 新增 SkillHubStorePanel 内嵌面板组件（替换原先的模态框）
- 修复图标加载失败时显示 fallback 图标
- 删除不再需要的 SkillHubModal.tsx
