+++ src/renderer/src/App.tsx
@@ -46,14 +46,12 @@ import {
   GitGraph,
   Minimize2,
   RefreshCw,
-  Package,
   X,
 } from "lucide-react";
 import { subscribeToNotice, showNotice } from "./utils/notice";
 import { createPreviewApi } from "./previewApi";
 import { createBrowserApi } from "./browserApi";
 const ConfigModal = lazy(() => import("./ConfigModal").then((m) => ({ default: m.ConfigModal })));
-const SkillHubModal = lazy(() => import("./components/app/SkillHubModal").then((m) => ({ default: m.SkillHubModal })));
 import { TrustConfirmModal } from "./components/app/TrustConfirmModal";
 import { TerminalDock } from "./components/terminal/TerminalDock";
 import { FeishuLinkIndicator } from "./components/feishu/FeishuLinkIndicator";
@@ -1049,7 +1047,6 @@ export function App() {
   const [sessionsProjectId, setSessionsProjectId] = useState<string>();
   const [sessionHistoryLoading, setSessionHistoryLoading] = useState(false);
   const [settingsOpen, setSettingsOpen] = useState(false);
-  const [skillHubOpen, setSkillHubOpen] = useState(false);
   const [updateInfo, setUpdateInfo] = useState<AppUpdateInfo | null>(null);
   const [updateError, setUpdateError] = useState<string | null>(null);
   const [updateChecking, setUpdateChecking] = useState(false);
@@ -6074,13 +6071,6 @@ export function App() {
                 <MessageSquare size={17} />
               </button>
               <button
-                className="icon-button skillhub-icon"
-                title={t("config.tabs.skillHub")}
-                onClick={() => setSkillHubOpen(true)}
-              >
-                <Package size={17} />
-              </button>
-              <button
                 className="icon-button homepage-icon"
                 title={t("app.homepage")}
                 onClick={() => api.app.openExternal("https://ayuayue.github.io/PiDeck/")}
@@ -7901,14 +7891,6 @@ export function App() {
         }}
       />
       </Suspense>
-
-      {skillHubOpen && (
-        <Suspense fallback={null}>
-          <SkillHubModal
-            onClose={() => setSkillHubOpen(false)}
-          />
-        </Suspense>
-      )}
 
       {confirmDialog && (
         <ConfirmDialog
