+++ src/renderer/src/config/SkillsTab.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Check, FileEdit, Pencil, ShoppingBag, ToggleLeft, ToggleRight, Trash2, X } from "lucide-react";
+import { Check, FileEdit, Pencil, ShoppingBag, ToggleLeft, ToggleRight, Trash2, X, Store } from "lucide-react";
 import type {
 	CreatePiSkillInput,
 	PiSkillListResult,
@@ -8,6 +8,7 @@ import type {
 } from "../../../shared/types";
 import { t } from "../i18n";
 import { SkillStoreTab } from "./SkillStoreTab";
+import { SkillHubStorePanel } from "./SkillHubStorePanel";
 
 export function SkillsTab(props: {
 	data: PiSkillListResult;
@@ -28,7 +29,7 @@ export function SkillsTab(props: {
 	onRename: (skill: PiSkillSummary, newName: string) => Promise<void>;
 }) {
 	const { data } = props;
-	const [skillTab, setSkillTab] = useState<"local" | "store">("local");
+	const [skillTab, setSkillTab] = useState<"local" | "store" | "skillhub">("local");
 	const [locationPickerOpen, setLocationPickerOpen] = useState(false);
 	const canCreate = props.newName.trim() && props.newDescription.trim();
 	// 按选中的位置目录过滤 skill 列表
@@ -53,6 +54,13 @@ export function SkillsTab(props: {
 					<ShoppingBag size={14} strokeWidth={1.8} />
 					{t("config.promptStoreTab")}
 				</button>
+				<button
+					className={`prompts-tab-btn ${skillTab === "skillhub" ? "active" : ""}`}
+					onClick={() => setSkillTab("skillhub")}
+				>
+					<Store size={14} strokeWidth={1.8} />
+					{t("config.tabs.skillHub")}
+				</button>
 			</div>
 
 			{skillTab === "store" ? (
@@ -60,6 +68,8 @@ export function SkillsTab(props: {
 					onImported={props.onRefresh}
 					locationId={props.newLocationId}
 				/>
+			) : skillTab === "skillhub" ? (
+				<SkillHubStorePanel />
 			) : (
 				<>
 					<div className="config-toolbar">
